refactor(errors): centralize error classification input
